% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{ask}
\alias{ask}
\title{NPPES Data Dissemination releases}
\usage{
ask(save = FALSE, path = fs::path_wd())
}
\arguments{
\item{save}{\verb{<lgl>} write data to disk with \code{data.table::fwrite()}; default is \code{FALSE}}

\item{path}{\verb{<chr>} path to save csv to; default is \code{fs::path_wd()}}
}
\value{
A \link[tibble:tibble-package]{tibble} of the search results.
}
\description{
NPPES Data Dissemination releases
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tmp <- fs::dir_create(fs::path_temp("nppez"))

nppez::ask(save = TRUE)

fs::dir_delete(tmp)
\dontshow{\}) # examplesIf}
}
