% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Retrieve info about the most recent NPPES Data Dissemination release}
\usage{
browse()
}
\value{
tibble with
}
\description{
Retrieve info about the most recent NPPES Data Dissemination release
}
\examples{
nppez::browse()
}
