% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{clean_credentials}
\alias{clean_credentials}
\title{Clean credentials}
\usage{
clean_credentials(x)
}
\arguments{
\item{x}{a string vector of provider credentials}
}
\value{
vector
}
\description{
Removes periods ("\code{.}") in strings
}
\examples{
clean_credentials("M.D.")

}
