% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_deactivation}
\alias{clean_deactivation}
\title{Monthly NPPES NPI Deactivation File}
\usage{
clean_deactivation(dir)
}
\arguments{
\item{dir}{\verb{<chr>} path to directory containing file}
}
\value{
tibble of cleaned data
}
\description{
Monthly NPPES NPI Deactivation File
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clean_deactivation("D:/<directory name>")
\dontshow{\}) # examplesIf}
}
