% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_endpoints}
\alias{clean_endpoints}
\title{Clean the Monthly NPPES NPI Endpoints File}
\usage{
clean_endpoints(csv = NULL)
}
\arguments{
\item{csv}{full path to Monthly NPPES Endpoints file (.csv)}
}
\value{
tibble of cleaned data
}
\description{
Clean the Monthly NPPES NPI Endpoints File
}
\examples{
\dontrun{
clean_endpoints("D:/<directory>/<filename>.csv")
}
}
