% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_locations}
\alias{clean_locations}
\title{Clean the Monthly NPPES NPI Secondary Practice Locations File}
\usage{
clean_locations(csv = NULL)
}
\arguments{
\item{csv}{full path to Monthly NPPES Secondary Practice Locations file (.csv)}
}
\value{
tibble of cleaned data
}
\description{
Clean the Monthly NPPES NPI Secondary Practice Locations File
}
\examples{
\dontrun{
clean_locations("D:/<directory>/<filename>.csv")
}
}
