% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_weekly_update}
\alias{clean_weekly_update}
\title{Clean the Weekly NPPES NPI Update File}
\usage{
clean_weekly_update(csv)
}
\arguments{
\item{csv}{full path to Weekly NPPES NPI Update (.csv) file}
}
\value{
tibble of cleaned data
}
\description{
Clean the Weekly NPPES NPI Update File
}
\examples{
\dontrun{
clean_weekly_update("D:/<directory>/<filename>.csv")
}
}
