% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{create_address}
\alias{create_address}
\title{Create Address Table from NPPES NPI File}
\usage{
create_address(df)
}
\arguments{
\item{df}{cleaned tibble from clean_nppes()}
}
\value{
tibble of cleaned data
}
\description{
Create Address Table from NPPES NPI File
}
\examples{
\dontrun{
create_address()
}
}
