% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{create_identifiers}
\alias{create_identifiers}
\title{Create Identifiers Table from NPPES NPI File}
\usage{
create_identifiers(df)
}
\arguments{
\item{df}{cleaned tibble from clean_nppes()}
}
\value{
tibble of cleaned data
}
\description{
Create Identifiers Table from NPPES NPI File
}
\examples{
\dontrun{
create_identifiers()
}
}
