% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{create_taxonomy}
\alias{create_taxonomy}
\title{Create Taxonomy Table from NPPES NPI File}
\usage{
create_taxonomy(df)
}
\arguments{
\item{df}{cleaned tibble from clean_nppes()}
}
\value{
tibble of cleaned data
}
\description{
Create Taxonomy Table from NPPES NPI File
}
\examples{
\dontrun{
create_taxonomy()
}
}
