% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{create_zip_file_names}
\alias{create_zip_file_names}
\title{Create Zip File Names for NBER NPI Datasets}
\usage{
create_zip_file_names(x)
}
\arguments{
\item{x}{a vector of file paths}
}
\value{
a vector of file names
}
\description{
Create Zip File Names for NBER NPI Datasets
}
\keyword{internal}
