% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{download_zips}
\alias{download_zips}
\title{Download NBER NPI Zip Files}
\usage{
download_zips(table, directory)
}
\arguments{
\item{table}{url to dataset's zip files}

\item{directory}{directory to save zip files}
}
\value{
results from \code{\link[curl:multi_download]{curl::multi_download()}} and a vector of zip file paths
}
\description{
Download NBER NPI Zip Files
}
\keyword{internal}
