% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{get_all_tables}
\alias{get_all_tables}
\title{Return All NBER NPI Dataset Information}
\usage{
get_all_tables()
}
\value{
data.frame
}
\description{
Return All NBER NPI Dataset Information
}
\keyword{internal}
