% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{get_table}
\alias{get_table}
\title{Get tibble of NBER NPI Dataset Information}
\usage{
get_table(url)
}
\arguments{
\item{url}{url to dataset's zip files}
}
\value{
data.frame
}
\description{
Get tibble of NBER NPI Dataset Information
}
\keyword{internal}
