% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{Split a Table by Groups with Named List Output}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{A data table or tibble.}

\item{...}{One or more unquoted variables by which to group and then split \code{.tbl}.
Variables can be separated by commas.}
}
\value{
A named list of tibbles/data tables, where each item corresponds to a unique combination
of the grouping variables. The names of the list items are derived from the unique
combinations of the grouping variables, separated by " / ".
}
\description{
This function takes a table, groups it by one or more variables, and then
splits the grouped data into a list. The resulting list has names derived
from the unique combinations of the grouping variables.
}
