% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{nber_urls}
\alias{nber_urls}
\title{NBER NPI Dataset URLs}
\usage{
nber_urls(x)
}
\arguments{
\item{x}{dataset group}
}
\value{
url to the dataset's zip files
}
\description{
NBER NPI Dataset URLs
}
\keyword{internal}
