% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppez-package.R
\docType{package}
\name{nppez-package}
\alias{nppez}
\alias{nppez-package}
\title{nppez: Tidy NPPES NPI Registry}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a tidy interface to the National Plan and Provider Enumeration System (NPPES) National Provider Identifier (NPI) Registry, a database of all health care providers who have been assigned an NPI. Includes functions to download and parse the NPI Registry data, perform updates according to the NPPES data release schedule and prepare it for storage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/nppez}
  \item \url{https://andrewallenbruce.github.io/nppez/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/nppez/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
