% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_seq}
\alias{rename_seq}
\title{Generate a sequence of numbers with a new prefix}
\usage{
rename_seq(
  n,
  new,
  old,
  between = " = ",
  enclose = NULL,
  collapse = ", ",
  style = TRUE
)
}
\arguments{
\item{n}{\verb{<int>} Number of sequences to generate}

\item{new}{\verb{<chr>} New prefix}

\item{old}{\verb{<chr>} Old prefix}

\item{between}{\verb{<chr>} Separator between \code{new} and \code{old}, default \code{" = "}}

\item{enclose}{\verb{<chr>} \emph{(optional)} Vector \code{length(x) == 2} with which to enclose output}

\item{collapse}{\verb{<chr>} Separator between sequences, default \code{", "}}

\item{style}{\verb{<lgl>} Apply \code{styler::style_text()} to output, default \code{TRUE}}
}
\value{
\verb{<chr>} collapsed vector of \code{n} sequences
}
\description{
Generate a sequence of numbers with a new prefix
}
\examples{
x <- rename_seq(
n        = 3, # can also be 25:300 etc.
new      = "id_issuer_",
old      = "Other.ID.Issuer.",
between  = " = ",
enclose  = c("x = c(", ")"),
collapse = ",\n ",
style    = TRUE)

}
