% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{single_line_string}
\alias{single_line_string}
\title{Format multiple line character vector to single line}
\usage{
single_line_string(x)
}
\arguments{
\item{`<chr>`}{character vector with line breaks (\verb{\\n})}
}
\value{
\verb{<chr>} single line character vector
}
\description{
Format multiple line character vector to single line
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
single_line_string("D:/")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
