% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{table_sum}
\alias{table_sum}
\title{Summarise tibble of NBER NPI Dataset Information}
\usage{
table_sum(table, directory)
}
\arguments{
\item{table}{url to dataset's zip files}

\item{directory}{directory to save zip files}
}
\value{
data.frame
}
\description{
Summarise tibble of NBER NPI Dataset Information
}
\keyword{internal}
