% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nber.R
\name{wrap}
\alias{wrap}
\title{Wrap strings}
\usage{
wrap(x, left, right)
}
\arguments{
\item{x}{a vector of provider credentials}

\item{left}{a string}

\item{right}{a string}
}
\value{
vector
}
\description{
Wrap strings
}
\examples{
wrap(x = "M.D.", left = "(", right = ")")

}
