% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{ask}
\alias{ask}
\title{List the most recent NPPES Data Dissemination releases}
\usage{
ask(save = FALSE, dir = NULL)
}
\arguments{
\item{save}{write to csv; default is \code{FALSE}}

\item{dir}{directory to save to}
}
\value{
tibble with
}
\description{
List the most recent NPPES Data Dissemination releases
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppez::ask()
\dontshow{\}) # examplesIf}
}
