% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{ask}
\alias{ask}
\title{List the most recent NPPES Data Dissemination releases}
\usage{
ask(save = FALSE, path = here::here("./"))
}
\arguments{
\item{save}{\verb{<chr>} write to disk using \code{readr::write_csv()} ; default is \code{FALSE}}

\item{path}{\verb{<chr>} path to save to; default is \code{here::here()}}
}
\value{
tibble with
}
\description{
List the most recent NPPES Data Dissemination releases
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppez::ask()
\dontshow{\}) # examplesIf}
}
