% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_deactivation}
\alias{clean_deactivation}
\title{Clean the Monthly NPPES NPI Deactivation File}
\usage{
clean_deactivation(dir_xlsx)
}
\arguments{
\item{dir_xlsx}{path to directory that the Monthly Deactivation Excel (.xlsx) file resides in (do not include file name at the end)}
}
\value{
tibble of cleaned data
}
\description{
Clean the Monthly NPPES NPI Deactivation File
}
\examples{
\dontrun{
clean_deactivation("D:/<directory>")
}
}
