% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_endpoints}
\alias{clean_endpoints}
\title{Monthly NPPES NPI Endpoints File}
\usage{
clean_endpoints(dir)
}
\arguments{
\item{dir}{\verb{<chr>} path to directory containing file}
}
\value{
tibble of cleaned data
}
\description{
Monthly NPPES NPI Endpoints File
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clean_endpoints("D:/<directory>/<filename>.csv")
\dontshow{\}) # examplesIf}
}
