% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{dispense}
\alias{dispense}
\title{Unzip NPPES ZIPs}
\usage{
dispense(zip_dir, unzip_dir)
}
\arguments{
\item{zip_dir}{path to directory containing ZIPs}

\item{unzip_dir}{path to directory to unzip ZIPs}
}
\value{
invisible
}
\description{
Unzip NPPES ZIPs
}
\examples{
\dontrun{
nppez::dispense()
}
}
