% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{dispense}
\alias{dispense}
\title{Unzip NPPES ZIPs}
\usage{
dispense(unzip, files = NULL)
}
\arguments{
\item{unzip}{path to directory to unzip ZIPs}

\item{files}{character vector of files inside a zip file to unzip}
}
\value{
invisible
}
\description{
Unzip NPPES ZIPs
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppez::dispense()
\dontshow{\}) # examplesIf}
}
