% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{grab}
\alias{grab}
\title{Download NPPES ZIP files to a local directory}
\usage{
grab(dir)
}
\arguments{
\item{dir}{path to local directory to download ZIPs to}
}
\value{
tibble
}
\description{
Download NPPES ZIP files to a local directory
}
\examples{
\dontrun{
nppez::grab(dir = tempdir())
}
}
