% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{grab}
\alias{grab}
\title{Download NPPES ZIP files to a local directory}
\usage{
grab(obj, files = NULL, path = fs::path_wd())
}
\arguments{
\item{obj}{\verb{<tbl_df>} object of class \code{nppez::ask}, returned from \code{nppez::ask()}}

\item{files}{\verb{<chr>} vector of files to download from ZIPs; default behavior is to download all files}

\item{path}{\verb{<chr>} path to download ZIPs to; default is \code{fs::path_wd()}}
}
\value{
tibble
}
\description{
Download NPPES ZIP files to a local directory
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppez::ask() |>
nppez::grab(files = "NPPES_Deactivated_NPI_Report_031124.zip")
\dontshow{\}) # examplesIf}
}
