% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{peek}
\alias{peek}
\title{Peek inside the downloaded NPPES ZIPs before unzipping}
\usage{
peek(path)
}
\arguments{
\item{path}{\verb{<chr>} path to download ZIPs to; default is \code{fs::path_wd()}}
}
\value{
tibble
}
\description{
Peek inside the downloaded NPPES ZIPs before unzipping
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppez::peek()
\dontshow{\}) # examplesIf}
}
