% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{peek}
\alias{peek}
\title{Peek inside the downloaded NPPES ZIPs before unzipping}
\usage{
peek(dir)
}
\arguments{
\item{dir}{path to directory of ZIPs}
}
\value{
tibble
}
\description{
Peek inside the downloaded NPPES ZIPs before unzipping
}
\examples{
\dontrun{
nppez::peek()
}
}
