% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{prune}
\alias{prune}
\title{Select files to unzip inside the downloaded NPPES ZIPs before unzipping}
\usage{
prune(dir)
}
\arguments{
\item{dir}{path to directory of ZIPs}
}
\value{
tibble
}
\description{
Select files to unzip inside the downloaded NPPES ZIPs before unzipping
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppez::prune()
\dontshow{\}) # examplesIf}
}
